Move TaskEditor styles into one file Fix a tag checkbox bug
