--- styles/scss/components/application/widgets/priorityPicker.module.scss
@@ -1,34 +1,0 @@
-@use "../../../utils/variable" as *;
-
-.popup_overlay {
-    position: fixed;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 1000;
-}
-
-.priority_picker {
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    background-color: #fff;
-    border-radius: 0.8rem;
-    box-shadow: 0px 1px 2rem 1px #c4c4c4;
-    overflow: hidden;
-    padding: 1rem 0;
-    z-index: 1001;
-}
-
-.priority_picker_button {
-    padding: .5rem .9rem;
-    display: flex;
-    justify-content: space-between;
-    font-size: 1.3rem;
-    cursor: pointer;
-    &:hover {
-        background-color: $grayColor;
-    }
-}
-
