--- styles/scss/components/application/widgets/scheduler.module.scss
@@ -1,132 +1,0 @@
-@use "../../../utils/variable" as *;
-
-
-.popup_overlay {
-    position: fixed;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 1000;
-}
-
-.scheduler {
-    position:absolute;
-    display: flex;
-    flex-direction: column;
-    background-color: #fff;
-    border-radius: 0.8rem;
-    box-shadow: 0px 1px 2rem 1px #c4c4c4;
-    overflow: hidden;
-    padding: 1rem 0;
-    z-index: 1001;
-}
-.scheduler_quickbutton{
-     padding:.5rem .9rem;
-     display: flex;
-     justify-content: space-between;
-     font-size: 1.3rem;
-     cursor: pointer;
- }
- .scheduler_quickbutton:hover{
-    background-color: $grayColor;
- }
-
-.datepicker_head {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    border-top: 1px solid $grayColor;
-    padding: 1rem .9rem .3rem;
-    font-size: 1.3rem;
-    font-weight: bold;
-}
-
-.datepicker_header_action{
-    button{
-        margin:0 .8rem;
-    }
-}
-
-.week_names {
-    display: flex;
-    color: $grayColor;
-    font-size: 1.2rem;
-    padding:.5rem .9rem;
-    border-bottom:1px solid $grayColor;
-    div{
-        width:3rem;
-        text-align: center;
-    }
-}
-
-
-
-.datepicker_monthlist_wrapper{
-    height: 19rem;
-    overflow: auto;
-    margin-right:-17px;
-}
-
-.datepicker_monthlist{
-    height:200000px;
-    position: relative;
-}
-
-.datepicker_calendar{
-    position: absolute;
-    left: 0;
-    right: 0;
-}
-.datepicker_calendar_header{
-    padding: .7rem 0;
-    margin:0 1.6rem;
-    border-bottom: 1px solid $grayColor;
-    font-size: 1.3rem;
-    font-weight: bold;
-}
-
-.datepicker_calendar_body{
-    display: flex;
-    flex-direction: column;
-}
-.datepicker_calendar_week {
-    display: flex;
-    flex-direction: row;
-    padding:.3rem .9rem;
-
-    button {
-        text-align: center;
-        width: 14.28571%;
-        line-height: 2rem;
-        font-size: 1.3rem;
-        cursor: pointer;
-    }
-    div{
-        width: 14.28571%;
-    }
-}
-.datepicker_selected{
-    display: block;
-    background-color: crimson;
-    border-radius: 50%;
-    color: #fff;
-    width: 2rem;
-    margin: 0 auto;
-    font-weight: bold;
-}
-.datepicker_normal{
-    display: block;
-    border-radius: 50%;
-    width: 2rem;
-    margin: 0 auto;
-    transition: all 100ms;
-}
-.datepicker_normal:hover{
-    background-color: $grayColor;
-}
-.datepicker_today{
-    font-weight: bold;
-    color: crimson;
-}
-
