+++ styles/scss/components/application/widgets/taskEditor.module.scss
@@ -12,7 +12,7 @@
 .task_edit_form{
     position: relative;
     background-color: #fff;
-    border:1px solid #e5e5e5;
+    border:1px solid $lightGrayColor;
     border-radius: 8px;
 }
 
@@ -29,7 +29,7 @@
     &:empty:before{
         content: attr(placeholder);
         position: absolute;
-        color: gray;
+        color: $textSecondaryColor;
         background-color: transparent;
     }
 }
@@ -45,8 +45,8 @@
     flex-direction: column;
     justify-content: flex-start;
     background-color: #fff;
-    border-radius: 5px;
-    box-shadow: 0 2px 10px rgba(0,0,0,.3);
+    border-radius: $borderRadius;
+    box-shadow: 0 2px 10px $grayColor;
     max-height: 15rem;
     overflow: auto;
     z-index: 1000;
@@ -54,12 +54,12 @@
         text-align: left;
         padding-left:1rem;
         &:hover{
-            background-color: #e5e5e5;
+            background-color: $grayColor;
         }
     }
 }
 .task_name_tag_selector_button_on{
-    background-color: #e5e5e5;
+    background-color: $grayColor;
 }
 
 .task_content{
@@ -70,7 +70,7 @@
     &:empty:before {
         content: attr(placeholder);
         position: absolute;
-        color: gray;
+        color: $textSecondaryColor;
         background-color: transparent;
     }
 }
@@ -78,7 +78,7 @@
 .task_edit_area{
     width: 100%;
     padding:1rem;
-    border-bottom:1px solid #e5e5e5;
+    border-bottom:1px solid $lightGrayColor;
 }
 
 .task_edit_buttons{
@@ -87,16 +87,174 @@
     justify-content: flex-start;
     button{
         border: 1px solid $grayColor;
-        border-radius: 4px;
+        border-radius: $borderRadiusSm;
         margin: 0 .3rem;
         padding: .5rem;
         font-size: 1.5rem;
         cursor: pointer;
         &:hover{
-            background-color: #e5e5e5;
+            background-color: $lightGrayColor;
         }
     }
 }
+
+/* Scheduler */
+.scheduler {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    background-color: #fff;
+    border-radius: $borderRadius;
+    box-shadow: 0px 1px 2rem 1px $grayColor;
+    overflow: hidden;
+    padding: 1rem 0;
+    z-index: 1001;
+}
+
+.scheduler_quickbutton {
+    padding: .5rem .9rem;
+    display: flex;
+    justify-content: space-between;
+    font-size: 1.3rem;
+    cursor: pointer;
+}
+
+.scheduler_quickbutton:hover {
+    background-color: $grayColor;
+}
+
+.datepicker_head {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    border-top: 1px solid $grayColor;
+    padding: 1rem .9rem .3rem;
+    font-size: 1.3rem;
+    font-weight: bold;
+}
+
+.datepicker_header_action {
+    button {
+        margin: 0 .8rem;
+    }
+}
+
+.week_names {
+    display: flex;
+    color: $grayColor;
+    font-size: 1.2rem;
+    padding: .5rem .9rem;
+    border-bottom: 1px solid $grayColor;
+
+    div {
+        width: 3rem;
+        text-align: center;
+    }
+}
+
+
+
+.datepicker_monthlist_wrapper {
+    height: 19rem;
+    overflow: auto;
+    margin-right: -17px;
+}
+
+.datepicker_monthlist {
+    height: 200000px;
+    position: relative;
+}
+
+.datepicker_calendar {
+    position: absolute;
+    left: 0;
+    right: 0;
+}
+
+.datepicker_calendar_header {
+    padding: .7rem 0;
+    margin: 0 1.6rem;
+    border-bottom: 1px solid $grayColor;
+    font-size: 1.3rem;
+    font-weight: bold;
+}
+
+.datepicker_calendar_body {
+    display: flex;
+    flex-direction: column;
+}
+
+.datepicker_calendar_week {
+    display: flex;
+    flex-direction: row;
+    padding: .3rem .9rem;
+
+    button {
+        text-align: center;
+        width: 14.28571%;
+        line-height: 2rem;
+        font-size: 1.3rem;
+        cursor: pointer;
+    }
+
+    div {
+        width: 14.28571%;
+    }
+}
+
+.datepicker_selected {
+    display: block;
+    background-color: $primaryColor;
+    border-radius: 50%;
+    color: #fff;
+    width: 2rem;
+    margin: 0 auto;
+    font-weight: bold;
+}
+
+.datepicker_normal {
+    display: block;
+    border-radius: 50%;
+    width: 2rem;
+    margin: 0 auto;
+    transition: all 100ms;
+}
+
+.datepicker_normal:hover {
+    background-color: $grayColor;
+}
+
+.datepicker_today {
+    font-weight: bold;
+    color: $primaryColor;
+}
+/* end Scheduler */
+
+/* priority picker */
+.priority_picker {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    background-color: #fff;
+    border-radius: $borderRadius;
+    box-shadow: 0px 1px 2rem 1px #c4c4c4;
+    overflow: hidden;
+    padding: 1rem 0;
+    z-index: 1001;
+}
+
+.priority_picker_button {
+    padding: .5rem .9rem;
+    display: flex;
+    justify-content: space-between;
+    font-size: 1.3rem;
+    cursor: pointer;
+
+    &:hover {
+        background-color: $grayColor;
+    }
+}
+/* end priority picker */
 
 
 .tag_check_list_dropdown{
@@ -105,8 +263,8 @@
     flex-direction: column;
     background-color: #fff;
     padding:1rem 0;
-    border-radius: 5px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, .3);
+    border-radius: $borderRadius;
+    box-shadow: 0 2px 10px $grayColor;
     max-height: 15rem;
     overflow: auto;
     z-index: 1001;
@@ -119,7 +277,7 @@
     padding:.25rem 1rem;
     cursor:pointer;
     &:hover{
-        background-color:#e5e5e5;
+        background-color:$grayColor;
     }
 }
 .tag_checkbox{
@@ -160,12 +318,12 @@
     background-color: #f5f5f5;
     cursor: pointer;
     &:hover {
-        background-color: #e5e5e5;
+        background-color: $lightGrayColor;
     }
 }
 .task_footer_submit{
     color: #fff;
-    background-color: #dc4c3e;
+    background-color: $primaryColor;
     cursor: pointer;
     &:hover {
         background-color: #c3392c;
